requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.12
langgraph==0.0.62
transformers==4.40.1
torch==2.7.0
accelerate==0.29.3
sentence-transformers==2.7.0
chromadb==0.4.24
requests==2.31.0
python-dotenv==1.0.1
PyYAML==6.0.1
